pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pg-reflinker"
version = "0.1.0"
description = "A boilerplate Python project with Kubernetes API client"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "kubernetes>=28.0.0",
    "kopf",
    "psycopg2-binary",
]

[project.scripts]
pg-reflinker = "pg_reflinker.main:run"

[tool.setuptools.packages.find]
where = ["src"]